Initial draft

+++ include/battle_ai_main.h
@@ -126,7 +126,7 @@ enum MoveComparisonResult
 
 void BattleAI_SetupItems(void);
 void BattleAI_SetupFlags(void);
-void BattleAI_SetupAIData(u8 defaultScoreMoves, u32 battler);
+void BattleAI_SetupAIData(u32 defaultScoreMoves, u32 battler);
 void ComputeBattlerDecisions(u32 battler);
 u32 BattleAI_ChooseMoveIndex(u32 battler);
 void Ai_InitPartyStruct(void);
